backend/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
redis==5.0.1
pyjwt==2.8.0
python-dotenv==1.0.0

